matches_service/web/requirements.txt
Django>=2.0
psycopg2
djangorestframework
markdown
django-filter
